Recommend the USE_GIT_REPOS var
As a first-time contributor to rspec, I was unfamiliar with the
dynamic Gemfile, and I ran into trouble because I had an old
clone sitting in just the right place.  The dynamic Gemfile
thought I wanted to use that old clone, so I got a dependency
conflict.

It seems to me that using dependencies from the filesystem is
a less common use-case than simply using the latest from git,
so I think the instructions should explicity USE_GIT_REPOS.