Demonstrate a bug with the output matcher.
An expectation of the form

```ruby
expect { ... }.to output.to_stderr
```

will work if anything is written to `$stderr` during the execution of the
block, but not if anything is written to `STDERR`. I encountered this
during working on a project and found the behaviour to be confusing.

Ideally both would work.