Refactor ExpectedsForMultipleDiffs#diffs to allocate less.
This also subtly changes the behaviour of the `diffs` method.  In the
old version, the diff would sometimes end with a newline, and sometimes
not. After the refactor, it never does. I had to make some changes
to tests to make the suite still pass; it is more consistent now.

Check for blank was shamelessly stolen from ActiveSupport's
`String#blank?`.