yield_with_args: match immediately
This change makes `yield_with_args(matcher)` evaluate the matcher as
soon as the arguments are yielded, rather than after the block has run.

This is more correct behaviour, as it allows testing of methods that
yield a value before making some change to it. Currently, there is no
way to use a yield matcher to test that a block yields a value with some
predicate being false, then returns with that predicate being true.