Handle sudden subject hash change
 #1034 introduced a hash comparison in addition to equality The change to compare not only the objects themselves, but also their
hashed has been introduced in #1034, and it broke the following examples
(using `rspec-rails`):

    it { expect { 'noop' }.not_to change { country.reload.cities } }

a workaround to that was to use the `.to_a` on the relation:

    it { expect { 'noop' }.not_to change { country.reload.cities.to_a } }

and the problem is with the objects that are changing their hash during
comparison.

Notable example:
https://github.com/rails/rails/blob/788eb51df3007d94b266965433b0be88a23b84c5/activerecord/lib/active_record/associations/collection_proxy.rb#L959

    def ==(other)
      load_target == other
    end

The problem is originally described here
https://github.com/rspec/rspec-expectations/issues/1030#issuecomment-450823945,
and amongst the proposed solution was to raise the question against
Rails team to fix the `CollectionProxy#hash` behaviour, but I believe
this wouldn't work for the following reasons:

1. The issue is specific to RSpec, while Rails comes with its own
testing framework that does the assertions using `==`, and is not
affected by the problem in question.

2. Fixing the issue would require to either:
 - stick to the original hash that the object was assigned that breaks
 the 'dynamic hashcode value depending on the contents of the object
 itself' that mutable Ruby objects seem to adhere to
 - pre-load the relation that may be a huge performance hit

The change swaps conditions in a way that the hashcode of the objects
being compared is not yet updated by the other sneaky condition.

This might also fix the issue described in #1070

The behaviour is described in documentation that probably needs to be
changed accordingly to this change.
http://rspec.info/documentation/3.8/rspec-expectations/RSpec/Matchers.html#change-instance_method

> Evaluates receiver.message or block before and after it evaluates the
> block passed to expect. If the value is the same object, its
> before/after hash value is used to see if it has changed. Therefore,
> your object needs to properly implement hash to work correctly with
> this matcher.