Prevent aliased matchers from overriding data
The aliased matcher implementation of a description currently
replaces all instance of the old matcher's name with the new
matcher's name. However, in cases where the expected value
has the same string as the old matcher's name, it would also
get replaced. This change makes it so only the first instance
of the old matcher's name is replaced in the description.