Improve composability of `include` matcher
The usage of hash keys used as arguments for `include` matcher was
limited by real keys, and it wasn't possible to pass in a matcher.

Ruby 1.8 behavior for `return` in `proc` is different from later
versions, e.g. if later version would use a value as a method return
value, Ruby 1.8 is not returning from a method. The behavior of
`Proc.new` is similar across Ruby versions.