swap ordering for values with units
The way `@delta * @expected.abs` is ordered currently, the expected
object *must* coerce into a Numeric to work with `#percent_of`.

Example of where this falls down:

```
expect(Weight.new("5 lb"))
  .to be_within(20).percent_of(Weight.new("4 lb"))
```

This is a statement that makes logical sense, but impossible to do with
`@delta * @expected.abs`.

By swapping the order, the expected merely has to respond to `#abs` and
`#*` and also be comparable.