Update StartWith, EndWith to use native methods if available
This addresses issue #1025.

With this change, the StartWith matcher will rely on an object's
start_with? method if available.  Similarly, the EndWith matcher
will rely on an object's end_with? method if available.

This is especially useful when a class implements start_with?
but not the indexing operator, or end_with? but not the indexing
operator.