Fail `Include` matcher when comparing String with non-String
If the actual is a String and the expected is not, a `TypeError` occurs
when we try to call `actual.include?`, becuase this method expects
a String when called on a String.

If the actual is a `String` and the expected is not, then the match
can't possibly succeed and we can return early.