nocov the AggregatedFailure message; update the spec
The message is out of date, since ruby 3 has gone _back_ to ignoring the
'inspect' output (this line was probably covered when it was written,
but version-specific, though I didn't bother pinning down exactly what
version). I updated the spec structure slightly as well, because the
original shape made it difficult to determine what the current message
actually _was_, but I'm happy to put it back if there's a style
preference.